UserIdの型の修正

+++ src/main/kotlin/com/ppap/torimocore/domain/User/User.kt
@@ -3,7 +3,7 @@ package com.ppap.torimocore.domain.User
 import java.io.Serializable
 import javax.persistence.*
 
-typealias UserId = Int
+typealias UserId = Long
 typealias ServiceId = String
 /**
  * ユーザー
